Details screen top sections

+++ app/src/main/java/com/loc/moviesfinder/core_feature/domain/model/MovieDetails.kt
@@ -3,7 +3,7 @@ package com.loc.moviesfinder.core_feature.domain.model
 data class MovieDetails(
     val id: Int,
     val title: String,
-    val averageVoting: Float,
+    val averageVoting: Double,
     val backdropPath: String?,
     val posterPath: String,
     val releaseYear: Int,
